Router Improvements
* Code cleanup
* Added ability to redirect after login
* Removed component loader from facebook view
