+++ frontend/src/components/FacebookAccountRedirect.vue
@@ -1,0 +1,120 @@
+<template>
+  <ComponentLoader
+    :loading-component="isLoadingComponent"
+    :failed-to-load="hasComponentFailedToLoad"
+  >
+    <div class="columns">
+      <div class="column">
+        <div v-if="fb_login_error != ''" class="level box is-marginless has-background-danger">
+          <span class="has-text-white">{{ fb_login_error | capitalize }}</span>
+        </div>
+      </div>
+    </div>
+  </ComponentLoader>
+</template>
+<script>
+import ComponentLoader from '@/components/ComponentLoader';
+import { capitalize } from '@/filters';
+
+export default {
+  name: 'FacebookCallback',
+  filters: {
+    capitalize,
+  },
+  components: { ComponentLoader },
+  props: {
+    formType: {
+      type: String,
+      default: '',
+    },
+    state: {
+      type: String,
+      default: '',
+    },
+    accessToken: {
+      type: String,
+      default: '',
+    },
+  },
+  data() {
+    return {
+      fb_login_error: '',
+    };
+  },
+  computed: {
+    redirect() {
+      try {
+        return JSON.parse(decodeURIComponent(this.state)).redirect;
+      } catch {
+        return '';
+      }
+    },
+  },
+  created() {
+    this.isLoadingComponent = true;
+    const vm = this;
+    if (typeof FB !== 'undefined' && window.location.hostname !== 'localhost') {
+      FB.getLoginStatus(function fbGetLogin(response) {
+        if (response.authResponse) {
+          vm.checkLoginState(response);
+        } else {
+          vm.$router.push({ path: '/login/' }).catch(() => {});
+        }
+      });
+    } else {
+      vm.dispatchLogin(this.accessToken);
+    }
+  },
+  methods: {
+    navigateAfterLogin() {
+      const nextRoute = this.redirect || '/my_profile/';
+
+      this.$router.push({ path: nextRoute }).catch(() => {
+        // the new version of vue router sends an error here
+        // so lets catch and do nothing
+      });
+    },
+    dispatchLogin(token) {
+      if (this.formType === 'register') {
+        this.$store
+          .dispatch('FB_REGISTER', { accessToken: token })
+          .then(() => {
+            this.navigateAfterLogin();
+          })
+          .catch((error) => {
+            this.fb_login_error = error.response.data.message || error;
+          })
+          .finally(() => {
+            this.isLoadingComponent = false;
+          });
+      } else {
+        this.$store
+          .dispatch('AUTH_FB_LOGIN', { accessToken: token })
+          .then(() => {
+            this.navigateAfterLogin();
+          })
+          .catch((error) => {
+            this.fb_login_error = error.response.data.message || error;
+          })
+          .finally(() => {
+            this.isLoadingComponent = false;
+          });
+      }
+    },
+    checkLoginState(response) {
+      if (response.status === 'connected') {
+        this.dispatchLogin(response.authResponse.accessToken);
+      } else if (response.status === 'not_authorized') {
+        // the user is logged in to Facebook,
+        // but has not authenticated your app
+        this.fb_login_error = 'Not authorized on Facebook';
+      } else {
+        // the user isn't logged in to Facebook.
+        this.fb_login_error = 'Facebook Error';
+      }
+    },
+  },
+};
+</script>
+
+<style scoped></style>
